// packages/typespec-test/test/Fleet.Management/generated/typespec-ts/src/models/models.ts
export interface ArmResource extends ArmResourceBase {
  /** Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName} */
  readonly id: string;
  /** The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts" */
  readonly type: string;
  /** Azure Resource Manager metadata containing createdBy and modifiedBy information. */
  readonly systemData?: SystemData;
}
// ...
export interface ProxyResourceBase extends ArmResource {}

/** An device resource belonging to a device group resource. */
export interface Device extends ProxyResourceBase {
  /** The resource-specific properties for this resource. */
  properties?: DeviceProperties;
}

/** The properties of device */
export interface DeviceProperties {
  /** Device ID */
  deviceId?: string;
  /** SKU of the chip */
  readonly chipSku?: string;
  /** OS version available for installation when update requested */
  readonly lastAvailableOsVersion?: string;
  /** OS version running on device when update requested */
  readonly lastInstalledOsVersion?: string;
  /** Time when update requested and new OS version available */
  readonly lastOsUpdateUtc?: Date;
  /** Time when update was last requested */
  readonly lastUpdateRequestUtc?: Date;
  /** The status of the last operation. */
  readonly provisioningState?: ProvisioningState;
}

/** Provisioning state of resource. */
/** "Succeeded", "Failed", "Canceled", "Provisioning", "Updating", "Deleting", "Accepted" */
export type ProvisioningState = string;
// ...
export interface DeviceUpdate {
  /** The updatable properties of the Device. */
  properties?: DeviceUpdateProperties;
}

/** The updatable properties of the Device. */
export interface DeviceUpdateProperties {
  /** Device group id */
  deviceGroupId?: string;
}
// ...
export interface DeviceListResult {
  /** The Device items on this page */
  value: Device[];
  /** The link to the next page of items */
  nextLink?: string;
}

/** Request of the action to create a signed device capability image */
export interface GenerateCapabilityImageRequest {
  /** List of capabilities to create */
  capabilities: CapabilityType[];
}

/** Capability image type */
/** "ApplicationDevelopment", "FieldServicing" */
export type CapabilityType = string;

/** Signed device capability image response */
export interface SignedCapabilityImageResponse {
  /** The signed device capability image as a UTF-8 encoded base 64 string. */
  readonly image?: string;
}